src/requirements.txt
iniconfig==2.0.0
markdown-it-py==2.2.0
mdurl==0.1.2
packaging==23.1
pluggy==1.0.0
Pygments==2.15.1
pyspellchecker==0.7.2
pytest==7.3.1
rich==13.3.5
